Actualizacion de migraciones

+++ database/migrations/2024_12_22_000905_add_user_id_to_pruebas_table.php
@@ -9,20 +9,27 @@ return new class extends Migration
     /**
      * Run the migrations.
      */
-    public function up()
+    public function up(): void
     {
-        Schema::table('pruebas', function (Blueprint $table) {
-            $table->unsignedBigInteger('user_id')->after('id'); // Agregar la columna user_id
-            $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade'); // Relación con la tabla users
-        });
+        if (!Schema::hasColumn('pruebas', 'user_id')) { // Verificar si la columna ya existe
+            Schema::table('pruebas', function (Blueprint $table) {
+                $table->unsignedBigInteger('user_id')->after('id'); // Agregar la columna user_id
+                $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade'); // Relación con la tabla users
+            });
+        }
     }
     
-    public function down()
+    /**
+     * Reverse the migrations.
+     */
+    public function down(): void
     {
-        Schema::table('pruebas', function (Blueprint $table) {
-            $table->dropForeign(['user_id']); // Eliminar la relación
-            $table->dropColumn('user_id'); // Eliminar la columna
-        });
+        if (Schema::hasColumn('pruebas', 'user_id')) { // Verificar si la columna existe antes de eliminarla
+            Schema::table('pruebas', function (Blueprint $table) {
+                $table->dropForeign(['user_id']); // Eliminar la relación
+                $table->dropColumn('user_id'); // Eliminar la columna
+            });
+        }
     }
-    
 };
+
